fix(popover): 调整demo
